refactor: change store to use redux toolkit
