+++ json-server/db.json
@@ -1,34 +1,16 @@
 {
   "events": [
     {
-      "id": 1,
-      "title": "Jogging",
-      "dateStart": "2021-08-11T08:18:09.629Z",
-      "dateEnd": "2021-08-11T09:18:09.629Z"
-    },
-    {
-      "title": "Teste 1",
-      "dateStart": "2021-11-09T16:06:06.488Z",
-      "dateEnd": "2021-11-09T16:06:08.499Z",
-      "id": 6
-    },
-    {
-      "title": "Teste 2",
-      "dateStart": "2021-11-09T17:47:28.231Z",
-      "dateEnd": "2021-11-09T17:47:31.043Z",
-      "id": 7
-    },
-    {
-      "title": "No nameasflsndflsdnf",
-      "dateStart": "2021-11-09T19:38:49.034Z",
-      "dateEnd": "2021-11-09T19:38:54.221Z",
-      "id": 8
+      "title": "No namedddd",
+      "dateStart": "2023-07-17T15:48:02.499Z",
+      "dateEnd": "2023-07-17T15:48:08.720Z",
+      "id": 1
     },
     {
       "title": "No name",
-      "dateStart": "2021-11-09T19:39:32.980Z",
-      "dateEnd": "2021-11-09T19:39:40.002Z",
-      "id": 9
+      "dateStart": "2023-07-18T16:39:10.186Z",
+      "dateEnd": "2023-07-18T16:39:15.600Z",
+      "id": 2
     }
   ]
 }
