+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import './index.css';
 import App from './components/App/App'
-import store from './components/redux/store';
+import store from './components/store/store';
 
 
 ReactDOM.render(
